refactor: make private
