+++ fglatch/__init__.py
@@ -1,0 +1,5 @@
+from fglatch._client.latch_client import LatchClient
+
+__all__ = [
+    "LatchClient",
+]
