+++ fglatch/_shared/type_aliases.py
@@ -2,8 +2,5 @@
 
 from typing import TypeAlias
 
-LatchUserApiToken: TypeAlias = str
-"""A Latch user API token."""
-
 LatchWorkspaceId: TypeAlias = str
 """A Latch workspace ID, e.g. "1234"."""
