--- fglatch/client/__init__.py
@@ -1,5 +1,0 @@
-from fglatch.client.latch_client import LatchClient
-
-__all__ = [
-    "LatchClient",
-]
